backend/requirements.txt
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
filelock==3.12.4
firebase-admin==6.2.0
Flask==3.0.0
fonttools==4.43.1
fsspec==2023.10.0
google-api-core==2.13.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
idna==3.4
imageio==2.31.6
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2
numpy==1.26.1
packaging==23.2
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-image==0.22.0
scipy==1.11.3
six==1.16.0
sympy==1.12
tifffile==2023.9.26
torch==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==3.0.1
